agent_package/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-team-orchestrator"
version = "0.3.3"
authors = [
  { name="Parth Bisht", email="parth@example.com" },
]
description = "A JIT Agent Factory that manifests multi-agent teams using Agno, LangGraph, and CrewAI on the fly."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "streamlit",
    "openai",
    "duckduckgo-search",
    "requests",
    "python-dotenv",
    "agno",
    "crewai",
    "langgraph"
]

[project.urls]
"Homepage" = "https://github.com/parthbisht/agentic-team-orchestrator"

[project.scripts]
agentic-factory = "agentic_team_orchestrator.cli:main"
